[CELEBORN-1259] Improve the default gracePeriod of ThreadUtils#shutdown
### What changes were proposed in this pull request?

Introduce `ThreadUtils#shutdown(executor)` method to improve the default gracePeriod of `ThreadUtils#shutdown`.

### Why are the changes needed?

The default value of `gracePeriod` for `ThreadUtils#shutdown` is 30 seconds at present. Meanwhile, the `gracePeriod` of most invoker for `ThreadUtils#shutdown` is 800 milliseconds. Therefore, the default `gracePeriod` of `ThreadUtils#shutdown` could be improved as 800 milliseconds.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No.

Closes #2276 from SteNicholas/CELEBORN-1259.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>