[CELEBORN-1241][FOLLOWUP] Fix duplicate CelebornRackResolver issue for SingleMasterMetaManager
### What changes were proposed in this pull request?
This pr followup CELEBORN-1241/https://github.com/apache/incubator-celeborn/pull/2246

For `SingleMasterMetaManager`, the given CelebornRackResolver is not used and a new one created in the constructor.

And in each CelebornRackResolver, there is `master-rack-resolver-refresher` thread pool. So, there is also duplicated thread pool issue.

### Why are the changes needed?
Fix duplicated `CelebornRackResolver` issue.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?
Not needed.

Closes #2315 from turboFei/rack_resolver.

Authored-by: Fei Wang <fwang12@ebay.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>