[CELEBORN-1288] Prompt configuration items when receiving IdleStateEvent
### What changes were proposed in this pull request?
When an `IdleStateEvent ` is received, the configuration items of the corresponding module are output.

### Why are the changes needed?
Now that the `IdleStateEvent` event is received, only the timeout time is output, but the corresponding configuration items are not output.

```
24/02/26 04:12:08,062 [data-client-5-8] ERROR TransportChannelHandler: Connection to /XXX:YYY has been quiet for 240000 ms while there are outstanding requests.
```

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
GA

Closes #2329 from cxzl25/CELEBORN-1288.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>
(cherry picked from commit f4bc87e6dbe6bb280d0e5771abad0e9353597c1e)
Signed-off-by: Cheng Pan <chengpan@apache.org>