[CELEBORN-1317][FOLLOWUP] Retry to setup mini cluster if the cause is BindException
To fix the UT for http server port already in use issue.

For Jetty HttpServer, if failed to bind port, the exception is IOException and the cause is BindException, we should retry for that.

Before:
```
    case e: BindException => // retry to setup mini cluster
```

Now:
```
    case e: IOException
         if e.isInstanceOf[BindException] || Option(e.getCause).exists(
           _.isInstanceOf[BindException]) =>  // retry to setup mini cluster
```

To fix the UT for http server port already in use issue.

No.

Will trigger GA for 3 three times.

Closes #2424 from turboFei/set_connector_stop_timeout.

Authored-by: Fei Wang <fwang12@ebay.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>