[CELEBORN-1351] Introduce SSLFactory and enable TLS support
### What changes were proposed in this pull request?

Add SSLFactory, and wire up TLS support with rest of Celeborn to enable secure over the wire communication.

### Why are the changes needed?
Add support for TLS to secure wire communication.
This is the last PR to add basic support for TLS.
There will be a follow up for CELEBORN-1356 and documentation ofcourse !

### Does this PR introduce _any_ user-facing change?
Yes, completes basic support for TLS in Celeborn.

### How was this patch tested?
Existing tests, augmented with additional unit tests.

Closes #2438 from mridulm/add-sslfactory-and-related-changes.

Authored-by: Mridul Muralidharan <mridulatgmail.com>
Signed-off-by: SteNicholas <programgeek@163.com>