[CELEBORN-1384] Manually excluding workers should not depend on whether the workers are alive
### What changes were proposed in this pull request?

Manually excluding workers should not depend on whether the workers are alive or not for master.

### Why are the changes needed?

When the workers are offline, master could not add or remove workers through manually excluding workers. Therefore, master should support manually excluding workers no matter whether the workers are alive or not.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manual test.

Closes #2465 from SteNicholas/CELEBORN-1384.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>