[CELEBORN-1318][FOLLOWUP] Transfer extraInfo for http authentication providers
### What changes were proposed in this pull request?
I am implementing the plugin for Bearer token authentication, and I found that, in ebay, the tokens are bound to client IP.

So, I also need to transfer the clientIp for token validation, I wonder that it is a general case.

This pr is a followup for Http password/token authentication and extend the current interface api.

### Why are the changes needed?
To extend the token authentication use case in case that we need more properties associate with the token.

### Does this PR introduce _any_ user-facing change?

No, this interface `TokenAuthenticationProvider` has not been released.

### How was this patch tested?

Not needed.

Closes #2604 from turboFei/auth_properties.

Lead-authored-by: Fei Wang <fwang12@ebay.com>
Co-authored-by: Cheng Pan <pan3793@gmail.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>