[CELEBORN-1541] Enhance the readable address for internal port
### What changes were proposed in this pull request?
If the internal port is not defined, do not show the internal port(-1) in the readable address.

### Why are the changes needed?

If the workerInfo is applied for the RESTful request, such as curl `/api/v1/workers/exclude`, the internal port is always `-1`.

It is not necessary to show the `-1` in the readable address.

### Does this PR introduce _any_ user-facing change?
Just reduce the unnecessary info.

### How was this patch tested?

Not needed.

Closes #2659 from turboFei/internal_port.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>
(cherry picked from commit dceef479bb217eb5f5b7dfa5728c3c05c4ac7e86)
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>