[CELEBORN-1535] Support to disable master workerUnavailableInfo expiration
### What changes were proposed in this pull request?

In this pr, it supports to disable the worker unavailable expiration by setting the timeout to -1.

### Why are the changes needed?

In our use case, we want to reserve all the worker unavailable information.
It is acceptable if we use the fixed ports and hosts, and will not occupy much memory resource.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?

Not needed.

Closes #2657 from turboFei/disable_Cleanup.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>
(cherry picked from commit a599ff2afe5847ebc3abdb53c8708747a5507bb4)
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>