[CELEBORN-1542] Master supports to check the worker host pattern on worker registration
### What changes were proposed in this pull request?
This pr introduce an optional config item for worker host pattern, and support to check whether the worker host matches the pattern in master end when registering the worker.

If it does not match, the register worker request will be rejected.

### Why are the changes needed?
Currently, the celeborn master allow all the workers to register. It is better to limit the workers allowed to register.

### Does this PR introduce _any_ user-facing change?

No, the config item is optional, no broken change.

### How was this patch tested?
UT.

Closes #2660 from turboFei/hosts_patterns.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>