[CELEBORN-1518] Add support for Apache Spark barrier stages
### What changes were proposed in this pull request?

Adds support for barrier stages.
This involves two aspects:
a) If there is a task failure when executing a barrier stage, all shuffle output for the stage attempt are discarded and ignored.
b) If there is a reexecution of a barrier stage (for ex, due to child stage getting a fetch failure), all shuffle output for the previous stage attempt are discarded and ignored.

This is similar to handling of indeterminate stages when `throwsFetchFailure` is `true`.

Note that this is supported only when `spark.celeborn.client.spark.fetch.throwsFetchFailure` is `true`

### Why are the changes needed?

As detailed in CELEBORN-1518, Celeborn currently does not support barrier stages; which is an essential functionality in Apache Spark which is widely in use by Spark users.
Enhancing Celeborn will allow its use for a wider set of Spark users.

### Does this PR introduce _any_ user-facing change?

Adds ability for Celeborn to support Apache Spark Barrier stages.

### How was this patch tested?

Existing tests, and additional tests (thanks to jiang13021 in #2609 - [see here](https://github.com/apache/celeborn/pull/2609/files#diff-e17f15fcca26ddfc412f0af159c784d72417b0f22598e1b1ebfcacd6d4c3ad35))

Closes #2639 from mridulm/fix-barrier-stage-reexecution.

Lead-authored-by: Mridul Muralidharan <mridul@gmail.com>
Co-authored-by: Mridul Muralidharan <mridulatgmail.com>
Signed-off-by: zky.zhoukeyong <zky.zhoukeyong@alibaba-inc.com>