[CELEBORN-1608] Reduce redundant response for RESTful api
### What changes were proposed in this pull request?
Currently, for exclude worker api, it will append all the workers info into the response message.
Assume that, we have thousands of workers, the response info is too large and redundant, as we already provide get workers api to return it.

### Why are the changes needed?

As title.

### Does this PR introduce _any_ user-facing change?
Yes, reduce the response.  It should be fine, we already provide the api to get workers details.

### How was this patch tested?
Existing UT.

Closes #2755 from turboFei/reduce_toomany_resp.

Authored-by: Wang, Fei <fwang12@ebay.com>
Signed-off-by: mingji <fengmingxiao.fmx@alibaba-inc.com>